Consolidate nested functions into separate functions Fixes #30
